requirements.txt
# Core dependencies
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.1.3

# PyTorch for embeddings
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Sentence Transformers
sentence-transformers==3.4.1

# FAISS for similarity search
faiss-cpu==1.7.4  # (Use faiss-gpu if running on GPU)

# OpenAI API for GPT chatbot
openai==1.61.1

# Streamlit for web app
streamlit==1.27.2

# Environment variable handling
python-dotenv==1.0.0
